backend/requirements.txt
fastapi
uvicorn
pillow
mangum
python-multipart